@@ -20,7 +20,8 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 13.0.1+27
 
 environment:
-  sdk: ">=2.18.2 <3.0.0"
+  sdk: ^3.0.0
+  flutter: ^3.10.0
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
